pyproject.toml
[project]
name = "vgg19-style-transfer"
version = "0.1.0"
description = "Notebook-backed VGG19 neural style transfer demo with progressive Gradio outputs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.3.0",
    "gdown>=5.2.0",
    "gradio>=5.23.0",
    "imageio[ffmpeg]>=2.37.0",
    "ipykernel>=6.30.1",
    "ipython>=9.6.0",
    "kagglehub>=0.3.13",
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.3",
    "pillow>=11.3.0",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.2",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "tqdm>=4.67.1",
    "wget>=3.2",
]
